Improved accuracy of room connections
The room connection algorithm assumed that edge tiles for a room were
approximately sorted, which was not the case. By sorting them, the
algorithm takes longer to run but produces significantly more accurate
results. Additionally, switched from the sup norm to euclidean norm, as
the performance penalty from the latter is not significant.