Cleanup pass and corrected height offset
Previously, the ceiling mesh would be generated at y = 0 and the walls
would extend down from there, which was unintuitive. Now the ground is
at y = 0 and the ceiling is raised to the wall height.

Mesh generator now operates at a higher level of abstraction, delegating
responsibilities to new lower level classes.

Removed multithreading from room connection algorithm as it's been
optimized to the point where this isn't necessary.