Cleaned up API
Initially, the intent was that the API would not use editor component at
all, which is not how things are normally done in Unity. Now, the values
in the editor component are used, and can be changed through the API.
This will allow users to set most values through inspector, then control
the ones that change through the API (like width/length).