Improved memory usage in triangulator
Originally, an array was being created for every square being
triangulated to store the vertices for that triangulation. Instead, a
single array is used for every square in a submap. This avoids having to
garbage collect tens or hundreds of thousands of small arrays which
could produce a frame rate spike after generating a map.