Improved texture quality on wall component
Initially, the vertex array for the walls was doubling up each vertex,
causing seems and walls with flat lighting. By removing this doubling
up, seams are eliminated and the walls are smoothly illuminated and
textured.

Additionally, changing the default shadowcasting settings on parts of
the cave dramatically reduces total vertex and triangle counts.