Implemented height maps
Formerly, terrain geometry was necessarily flat, relying on textures to
create the illusion of depth. Now, it's possible to use height maps to
create actual depth throughout the terrain. This dramatically improves
the overall look and feel of generated terrain.

Also reorganized files, as the number of class files is growing large.