Refactored mesh generator and mesh builder classes
Instead of requiring the cave generator to configure the mesh generator,
the mesh generator offers methods that handle the configuration process.
Also moved around some of the responsibilities in the various mesh
builders. This improved performance of enclosure caves and fixed a bug
where an unnecessary extra wall around the entire cave was being
generated.