Isolated physics calculations
The PhysX engine does a lot of calculations when first activating a mesh
with colliders. Initially these were interspersed with the cave
generator. By disabling game objects until the generator is finished,
the calculations are all done at the end. The improves responsiveness
and will maker it easier to add logic later to enable only some of the
meshes at the start.