Made floodfill algorithms in MapGeneration more robust
Originally, the floodfill functionality operated on the assumption that
the outer boundary would always consist of walls. This allowed for a
clever optimization that removed the need for boundary checks during BFS
routines. To accommodate maps not satisfying this assumption, boundary
checks have been restored and the optimization removed. This results in
a minor performance penalty.