Optimized marching squares algorithm
Reduced the run time of the map triangulator by more than half, and
simplified the vertex caching. The main change was to do two passes
instead of one. The first pass computed the configuration for each
square, then the second triangulated based on configuration. This
simplified the algorithm and revealed some significant optimizations.