Complete redesign of outline generator
The outline generator is now 5x faster and uses half the memory. In the
process, several classes designed to address issues with the old
generator were removed. The new one exploits the structure of the
marching squares algorithm, whereas the old one used mesh geometry.

This has the side effect of fixing a long-standing and tricky issue with
the textures of walls not matching up across cave chunks.