Fixed prefab creation bug (Fixes #2)
The code to create a cave prefab first created the prefab, then
attempted to access the meshes on the prefab to serialize those as
assets, but was unable to find them resulting in an
InvalidOperationException. This has been fixed by getting the meshes
from the in-memory object instead. Credit to camta005 for reporting the
bug.