Redesigned MapBuilder
MapBuilder originally provided extension functions written in pure
functional style (no side effects). This was undesirable for two
reasons. One is that it was inconsistent with the non-functional methods
that already existed on Map. Two is that it created a lot of allocations
on the Large Object Heap (LOH) for large maps. Rewritten to instead
provide static functions that mutate existing maps where possible.