Implemented new cave type
Implemented a new type of cave that generates a floor in the usual way,
but instead of walls and a ceiling, instantiates rock prefabs along the
outlines. A new OutlineModule module type specifies how the outlines are
processed. The defaut module lets prefabs be assigned with weightings
(probabilities of instantiation). Numerous changes across many types
(mostly refactorings) were made to support the new type.