Overall cleanup. Improved map widening algorithm.
The map widening algorithm was messy - it was very aggressive for large
values and had some edge cases it didn't handle properly. Replaced it
with a simple algorithm that only handles a value of 1, but handles it
properly.

Also cleaned up the readmes and improved the way entrances/boundaries
are handled in the map modules.