Refactored CaveGenerator and its editors
Replaced CaveGenerator's ad hoc design with a simple inheritance tree,
and created a factory class to abstract away the instantiation details.
Similarly replaced the single unmaintanable editor class
CaveGeneratorUIEditor with a facade that delegates to an editor for the
type of cave generator currently being inspected.

These designs will be easier to maintain and extend, and will facilitate
future object pooling.